requirements.txt
flask
python-docx
google-generativeai
python-dotenv
protobuf==4.25.3
